sync and async pipelines functional, still need to update tests
