+++ prime-router/src/main/resources/db/migration/V23__add_process_type.sql
@@ -1,0 +1,15 @@
+/*
+ * The Flyway tool applies this migration to create the database.
+ *
+ * Follow this style guide https://about.gitlab.com/handbook/business-ops/data-team/platform/sql-style-guide/
+ * use VARCHAR(63) for names in organization and schema
+ *
+ * Copy a version of this comment into the next migration
+ *
+ */
+
+/*
+ * Add process enum type to underlying custom data type
+ */
+ALTER TYPE public.task_action ADD VALUE 'process' after 'receive'
+;
