+++ prime-router/src/test/kotlin/azure/SendFunctionTests.kt
@@ -49,6 +49,7 @@ class SendFunctionTests {
         null,
         null,
         null,
+        null,
         null
     )
     private val reportFile = ReportFile(
